Update Multi Sync Env Runner and test with PPO on lunar lander
