+++ ai_traineree/__init__.py
@@ -1,6 +1,6 @@
 import torch
 
-__version__ = "0.4.2"
+__version__ = "0.4.3"
 
 
 try:
