+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
-    "setuptools>=42",
-    "wheel"
+  "setuptools>=42",
+  "wheel",
 ]
 
 build-backend = "setuptools.build_meta"
